backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jonatan Sotelo <bbip@live.com.ar>"]
readme = "README.md"
packages = [{ include = "app" }]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.28.1"
pytest-timeout = "^2.3.1"
# opcional si querés usar @pytest.mark.order
pytest-order = "^1.3.0"


# Web
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }

# Config/validación
python-dotenv = "^1.1.1"
pydantic = { version = "^2.11.7", extras = ["email"] }
pydantic-settings = "^2.10.1"

# Logging
loguru = "^0.7.3"

# DB / ORM / Migrations
sqlalchemy = "^2.0.41"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"

# Auth/JWT
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
python-multipart = "^0.0.9"

# Excel export
openpyxl = "^3.1.5"

# Scheduler
APScheduler = "^3.10.4"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
